assume principal id is always a number (#7)
This is a follow on PR to PR #1 to address the code used to check
whether the principal id is a number or an email. Since we are only
allowing access thru synapse IDP we can safely assume that the
principal ID is always going to be the synapse ID nunber and not an
email address.  This refactor removes that check.