[SC-168] product name tag and access approved role tag
We currently add the product name tag and access approved role tag
to EC2 instances using a bash script[1].  The idea here is to
replace the those scripts by doing the same thing with the
synapse tagger.  Using the synapse tagger makes it easier
to use and maintain going forward.  Also it would be good to
consolidate all tagging actions to the synapse tagger.

[1] https://github.com/Sage-Bionetworks/service-catalog-utils/blob/master/linux/opt/sage/bin/apply_name_tag.sh