[SC-168] Tag attached EC2 volumes
We currently tag volumes that are attached to provisioned EC2 instance
with scripts inside of CFN template files[1].  The idea with this change
is to move that functionality to the Synapse tagger. This will consolidate
resource tagging to this lambda and make it easier to apply and maintain
going forward.

Also there's a slight refactor the handlers to make it easier to test.

[1] https://github.com/Sage-Bionetworks/service-catalog-library/blob/master/ec2/sc-ec2-linux-jumpcloud.yaml