[SC-248] Exclude synapse userProfile data (#18)
* [SC-248] Exclude synapse userProfile data

S3 buckets do not allow tag values with certain characters, like a
comma (i.e. Seattle, WA).  This change excludes additional userProfile data[1]
that may contain restricted characters.

The only doc i've run across that touches on character restrictions for tag
values is in the EC2 docs[2]

"Although EC2 allows for any character in its tags, other services are more
restrictive. The allowed characters across services are: letters, numbers,
and spaces representable in UTF-8, and the following characters: + - = . _ : / @."

So it looks like it's not consistent across services. Commas are allowed for
EC2s tags but not for S3 tags.

[1] http://rest-docs.synapse.org/rest/org/sagebionetworks/repo/model/UserProfile.html
[2] https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_Tags.html#tag-restrictions

* include synapse user data instead of exclude

This change changes the logic from exclusion to inclusion of synapse
userProfile data as tags.

* sanitize tag values