Skip applying tags if ENV VAR is not defined. (#23)
We had to revert PR https://github.com/Sage-Bionetworks/scipool-infra/pull/225
because this lambda would fail to run if dependent ENV VARS are not setup.
For example the `MarketplaceProductCodeSC` has not been setup in production
environments because we have not completed development and testing of
the aws marketplace integration in the develop environment yet.

This change makes it so that the labmda doesn't error when the ENV VARs
are not setup.  Instead the lambda will just skip applying tags if the
required ENV VAR is not defined.  This will allow us to deploy the
lambda in all environments.