[SC-379] Refactor set batch tags
The current implementation of the SetBatchTagsFunction does not work due
to a bug in AWS where the Service Catalog does not apply SC tags to batch
resources.  The SC tags do appear on cloudformation stacks so we
refactor the function to get the SC principal ARN from the
cloudformation stacks instead of the batch resources.