Update dependencies
A recent PR proposed by dependabot is failing to install dependencies
in TravisCI. Update dependencies using `pipenv update --dev` to resolve
the security alert, and update to python3.10 to avoid a conflict
between `setuptools` and `importlib-metadata`.

Also upgrade the OS to Bionic for an approximate kernel match with the
lambda runtime environment.

https://askubuntu.com/questions/517136/list-of-ubuntu-versions-with-corresponding-linux-kernel-version
https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html