[IT-4483] Refactor get_synapse_owner_id (#63)
Refactor get_synapse_owner_id method to get the owner ID from both synapse::ownerId and aws:servicecatalog:provisioningPrincipalArn tags with the former as the preferred option and the latter as the fallback option.

This implements the following proposal in IT-4483:
```
For a SC product Updates action the lambda currently also gets the synapse id
from the aws:servicecatalog:provisioningPrincipalArn tag.  We would need to
refactor the lambda to get the synapse id from thesynapse:owerId tag instead.
Then the lambda will generate and apply all of the synapse prefixed tags with the
new owner info. 
```